PPTX to PDF now is partially implemented.
ChatGPT hallucinated that python-pptx had a function that could export a pptx slide to an image
(or anything other than pptx for that record). This was researched and deemed a falsification.

Instead, we're now using LibreOffice to do that - it dumps the entire pptx to a pdf,
which we then reimport into our program.

Theoretically, python-pptx has some functions that might be able to detect if a video is on, say,
slide 2, but it is unclear if python-pptx can export that video.

Issues:
LibreOffice is not expecting to export a subsection of a pptx file, which will be
an issue when we finally attempt to line up outputted png files and videos using file names.
(ex: slide1.png, slide2.mp4, slide3.png, slide4.png)